Refactor to move sidecar manager side-functions and tests over which to better homes.
